+++ packages/app/app-template/.storybook/mocks/next-headers.ts
@@ -1,0 +1,13 @@
+export function headers() {
+	return new Headers();
+}
+
+export function cookies() {
+	return {
+		get: () => undefined,
+		getAll: () => [],
+		set: () => {},
+		delete: () => {},
+		has: () => false,
+	};
+}