+++ packages/app/app-template/.storybook/mocks/next-link.tsx
@@ -1,0 +1,19 @@
+import { forwardRef } from "react";
+
+type LinkProps = {
+	href: string;
+	children?: React.ReactNode;
+	[key: string]: unknown;
+};
+
+const Link = forwardRef<HTMLAnchorElement, LinkProps>(
+	({ href, children, ...props }, ref) => (
+		<a href={href} ref={ref} {...props}>
+			{children}
+		</a>
+	),
+);
+
+Link.displayName = "Link";
+
+export default Link;